+++ src/main/java/com/example/outsourcing/global/exception/EmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.example.outsourcing.global.exception;
+
+//이미 사용중인 이메일임
+public class EmailAlreadyExistsException extends RuntimeException {
+    public EmailAlreadyExistsException() {
+    }
+}
