+++ src/main/java/com/example/outsourcing/global/exception/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
 package com.example.outsourcing.global.exception;
 
-import com.example.outsourcing.global.exception.Exceptions.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
