+++ src/main/java/com/example/outsourcing/global/exception/PasswordMissingDigitException.java
@@ -1,0 +1,7 @@
+package com.example.outsourcing.global.exception;
+
+//숫자가 1개 이상 포함되어야함
+public class PasswordMissingDigitException extends RuntimeException {
+    public PasswordMissingDigitException() {
+    }
+}
