+++ src/main/java/com/example/outsourcing/global/exception/PasswordTooShortException.java
@@ -1,0 +1,7 @@
+package com.example.outsourcing.global.exception;
+
+//비밀번호가 너무 짧음
+public class PasswordTooShortException extends RuntimeException {
+    public PasswordTooShortException() {
+    }
+}
