+++ src/main/java/com/example/outsourcing/service/UserService.java
@@ -6,8 +6,8 @@ import com.example.outsourcing.dto.LoginRequestDto;
 import com.example.outsourcing.config.PasswordEncoder;
 import com.example.outsourcing.dto.UserCreateRequestDto;
 import com.example.outsourcing.dto.UserCreateResponseDto;
+import com.example.outsourcing.global.exception.*;
 import com.example.outsourcing.repository.UserRepository;
-import com.example.outsourcing.global.exception.Exceptions.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
